requirements.txt
# F1 Historical Dataset - Python Dependencies

# Core dependencies
requests>=2.31.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0

# FastF1 library for modern F1 data
fastf1>=3.1.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# PDF parsing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Database
# SQLite is included in Python standard library
# For PostgreSQL: psycopg2-binary>=2.9.0

# Data export
pyarrow>=14.0.0  # For Parquet export

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Ergast API client (optional, we'll use requests directly)
# ergast-python>=0.1.0

# Rate limiting
ratelimit>=2.2.1

# Logging and progress
tqdm>=4.66.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

